fixed signed/unsigned comparisons

+++ src/lib/libc/stdlib/dtoa.c
@@ -424,7 +424,7 @@ _DEFUN (_dtoa_r,
 	i = 1;
     }
   j = sizeof (__ULong);
-  for (_REENT_MP_RESULT_K(ptr) = 0; sizeof (_Bigint) - sizeof (__ULong) + j <= i;
+  for (_REENT_MP_RESULT_K(ptr) = 0; sizeof (_Bigint) - sizeof (__ULong) + j <= (size_t)i;
        j <<= 1)
     _REENT_MP_RESULT_K(ptr)++;
   _REENT_MP_RESULT(ptr) = Balloc (ptr, _REENT_MP_RESULT_K(ptr));
